server/app/requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.3
importlib-metadata==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
packaging==23.0
pluggy==1.0.0
PyMySQL==1.0.2
pytest==7.2.2
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
Flask-Cors==3.0.10